Create recipe pages
Three recipe pages were made, and the information on each page is organized
in the same layout. There is a title for the recipe, a picture of the food
after being made, a description of the recipe, a list of ingredients, and a
list of steps to be taken to make the food.